+++ test/Reader.ts
@@ -128,4 +128,10 @@ describe('Reader', () => {
   it('sequenceArray', () => {
     U.deepStrictEqual(pipe([_.of(1), _.of(2)], _.sequenceArray)(undefined), [1, 2])
   })
+
+  it('asksE', () => {
+    const e: Env = { count: 0 }
+    const f = (e: Env) => _.of(e.count + 1)
+    U.deepStrictEqual(_.asksE(f)(e), 1)
+  })
 })